# src/openapi.yaml
openapi: 3.0.1
info:
  title: User API
  description: The users API description
  version: '1.0'
paths:
  /users:
    post:
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    get:
      summary: Get all users
      operationId: getAllUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  '/users/{id}':
    get:
      summary: Get a user by ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update a user by ID
      operationId: updateUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete a user by ID
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
components:
  schemas:
    CreateUserDto:
      type: object
      required:
        - name
        - email
        - birthDate
        - role
      properties:
        name:
          type: string
        email:
          type: string
        birthDate:
          type: string
          format: date
        role:
          type: string
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        birthDate:
          type: string
          format: date
        role:
          type: string
    User:
      type: object
      required:
        - id
        - name
        - email
        - birthDate
        - role
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        birthDate:
          type: string
          format: date
        role:
          type: string